moving common auth items to identity.ts and fixing UserRecord definitions
