+++ spec/common/providers/identity.spec.ts
@@ -1,0 +1,88 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2017 Firebase
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+
+import { expect } from 'chai';
+import * as identity from '../../../src/common/providers/identity';
+
+describe('identity', () => {
+  describe('userRecordConstructor', () => {
+    it('will provide falsey values for fields that are not in raw wire data', () => {
+      const record = identity.userRecordConstructor({ uid: '123' });
+      expect(record.toJSON()).to.deep.equal({
+        uid: '123',
+        email: null,
+        emailVerified: false,
+        displayName: null,
+        photoURL: null,
+        phoneNumber: null,
+        disabled: false,
+        providerData: [],
+        customClaims: {},
+        passwordSalt: null,
+        passwordHash: null,
+        tokensValidAfterTime: null,
+        metadata: {
+          creationTime: null,
+          lastSignInTime: null,
+        },
+      });
+    });
+
+    it('will not interfere with fields that are in raw wire data', () => {
+      const raw: any = {
+        uid: '123',
+        email: 'email@gmail.com',
+        emailVerified: true,
+        displayName: 'User',
+        photoURL: 'url',
+        phoneNumber: '1233332222',
+        disabled: true,
+        providerData: [],
+        customClaims: {},
+        passwordSalt: 'abc',
+        passwordHash: 'def',
+        tokensValidAfterTime: '2027-02-02T23:01:19.797Z',
+        metadata: {
+          creationTime: '2017-02-02T23:06:26.124Z',
+          lastSignInTime: '2017-02-02T23:01:19.797Z',
+        },
+      };
+      const record = identity.userRecordConstructor(raw);
+      expect(record.toJSON()).to.deep.equal(raw);
+    });
+
+    it('will convert raw wire fields createdAt and lastSignedInAt to creationTime and lastSignInTime', () => {
+      const raw: any = {
+        uid: '123',
+        metadata: {
+          createdAt: '2017-02-02T23:06:26.124Z',
+          lastSignedInAt: '2017-02-02T23:01:19.797Z',
+        },
+      };
+      const record = identity.userRecordConstructor(raw);
+      expect(record.metadata).to.deep.equal({
+        creationTime: '2017-02-02T23:06:26.124Z',
+        lastSignInTime: '2017-02-02T23:01:19.797Z',
+      });
+    });
+  });
+});
