+++ src/providers/auth.ts
@@ -20,8 +20,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-import * as firebase from 'firebase-admin';
-import * as _ from 'lodash';
+import {
+  UserRecord,
+  userRecordConstructor,
+} from '../common/providers/identity';
 import {
   CloudFunction,
   Event,
@@ -52,21 +54,9 @@ export function _userWithOptions(options: DeploymentOptions) {
   }, options);
 }
 
-export class UserRecordMetadata implements firebase.auth.UserMetadata {
-  constructor(public creationTime: string, public lastSignInTime: string) {}
-
-  /** Returns a plain JavaScript object with the properties of UserRecordMetadata. */
-  toJSON() {
-    return {
-      creationTime: this.creationTime,
-      lastSignInTime: this.lastSignInTime,
-    };
-  }
-}
-
 /** Builder used to create Cloud Functions for Firebase Auth user lifecycle events. */
 export class UserBuilder {
-  private static dataConstructor(raw: Event): firebase.auth.UserRecord {
+  private static dataConstructor(raw: Event): UserRecord {
     return userRecordConstructor(raw.data);
   }
 
@@ -108,73 +98,4 @@ export class UserBuilder {
       options: this.options,
     });
   }
-}
-
-/**
- * The UserRecord passed to Cloud Functions is the same UserRecord that is returned by the Firebase Admin
- * SDK.
- */
-export type UserRecord = firebase.auth.UserRecord;
-
-/**
- * UserInfo that is part of the UserRecord
- */
-export type UserInfo = firebase.auth.UserInfo;
-
-export function userRecordConstructor(
-  wireData: Object
-): firebase.auth.UserRecord {
-  // Falsey values from the wire format proto get lost when converted to JSON, this adds them back.
-  const falseyValues: any = {
-    email: null,
-    emailVerified: false,
-    displayName: null,
-    photoURL: null,
-    phoneNumber: null,
-    disabled: false,
-    providerData: [],
-    customClaims: {},
-    passwordSalt: null,
-    passwordHash: null,
-    tokensValidAfterTime: null,
-  };
-  const record = _.assign({}, falseyValues, wireData);
-
-  const meta = _.get(record, 'metadata');
-  if (meta) {
-    _.set(
-      record,
-      'metadata',
-      new UserRecordMetadata(
-        meta.createdAt || meta.creationTime,
-        meta.lastSignedInAt || meta.lastSignInTime
-      )
-    );
-  } else {
-    _.set(record, 'metadata', new UserRecordMetadata(null, null));
-  }
-  _.forEach(record.providerData, (entry) => {
-    _.set(entry, 'toJSON', () => {
-      return entry;
-    });
-  });
-  _.set(record, 'toJSON', () => {
-    const json: any = _.pick(record, [
-      'uid',
-      'email',
-      'emailVerified',
-      'displayName',
-      'photoURL',
-      'phoneNumber',
-      'disabled',
-      'passwordHash',
-      'passwordSalt',
-      'tokensValidAfterTime',
-    ]);
-    json.metadata = _.get(record, 'metadata').toJSON();
-    json.customClaims = _.cloneDeep(record.customClaims);
-    json.providerData = _.map(record.providerData, (entry) => entry.toJSON());
-    return json;
-  });
-  return record as firebase.auth.UserRecord;
 }
